moving labs API iiif catalog to prod service (#67)
Uses the scrape API to produce results for `/` and `/collection.json` endpoints.

By default, if no query is provided, the `restrict_to_iiif` collections query is used which is `(mediatype:'texts' AND mediatype:'image')`. Additionally, `restrict_to_iiif` can be used to limit any other query to these collections (in attempt to produce results that are IIIF compatible). In the future, this should use `formats:` instead of `mediatype:` to include AV and produce a more complete and accurate set of compatible items.


---------

Co-authored-by: Glen Robson <glenrobson@users.noreply.github.com>